Fix a bug which caused naming of columns to be critical to hash production
Prior to this change, if two columns in a given dataset query had the same name, the hash value could be indeterminate. Prior implementation used `row.get(colByName)` to lookup the col, however this does not fail on duplicate and leads to indeterminate results. Instead the implementation can rely on the ordering of `columnMetadatas` and do positional lookups of each column.

This makes the ordering critical, as it was intended to be.