Bump Java from 16.0.2+7 to 17.0.0+35 (LTS) Bump Kotlin plugins from 1.5.21 to 1.6.0-RC2 Bump Kotlin from 1.5.31 to 1.6.0-RC2
There were issues previously with Kotlin Gradle Kapt plugin under JVM 16 which needed a workaround, even worse workarounds on JVM 17, and issues with the latest 1.5.31 release of the plugins. Kotlin 1.6 seems to address these issues and allows us to align the Kotlin plugin versions and Kotlin runtime versions which is better practice.

Not ideal to rely on a release candidate, and Micronaut is not officially tested with it in its BOM but it should be released soon, and it cleans things up.