Force set the JUnit BOM version to avoid breaking running unit tests on IntelliJ IDEA
There seems to be an issue with IDEA detecting the correct junit platform version to run tests with after the upgrade to Mockito 4.0.0, which pulled in JUnit platform 1.8 (rather than 1.7). It seems IntelliJ doesn't play nicely with the way Micronaut is configuring the BOM versions for JUnit, and it can't detect the right version to run with. Setting the JUnit platform explicitly seems to workaround this.