#40 Correct batch loading of data with large batch sizes by limiting the concurrency of `flatMap`
Previously thousands of elements were streaming out of the batches and causing parallel attempts to get DB connections. When combined with updating existing rows, this was causing connection exhaustion and actually deadlocks; I think because different batches were creating a connection to bulk look for existing rows, and then then updating those rows one-by-one, which could lead to deadlock if there was no connection available for them to be able to proceed.

Suspect we still need to do more here (limit concurrency on the update rows flatMap?) or otherwise it's still possible for different rec runs in parallel to deadlock on each other due to connection starvation.