#66 Change hashing implementation to reduce change of false matches
Currently if there are two adjacent string columns with values "abc" and "def" vs "ab" and "cdef" the rows would be considered equal via the hashing function. This introduces a delimiter between values to reduce the chance of such false positives by inserting a separator between field values as they are hashed.