#39 Introduce HashingStrategy to allow for more lenient hashing w.r.t types
In practice, different drivers return different types at runtime. The previous strategy is strict, i.e if one driver/DB column returns a 32-bit int and another returns a 64-bit long, they will be considered unequal, regardless of their values. This is probably not desirable. This introduces a strategy that allows a more lenient approach to hashing, which is most likely to be desirable, especially when comparing across DB types, and different schemas which might change the types at DB level for efficiency.