Move mysql to the start of the database list
There appears to be some race condition or indeterminism when loading data into SQL Server that can cause it to have no rows in the TestData table when the rec is run. This usually only happens for a single test case, and not always, and I cannot replicate locally. Trying this to see if the problem still exists when sql server is always the source DB rather than the target.