#155 Workaround to ensure the ReactorDebugAgent works OK when using Gradle to run under colima
It's not entirely clear why, but when launching via Gradle / Micronaut Application Run plugin under colima (Docker mode) the JVM doesn't seem to be able to use its dynamic attachment mechanism out-of-the-box to connect. This probably doesn't seem to occur when running an "officially built" Recce image under colima so seems related to either batect or the Gradle spawn mechanism. Neither "external" attach or "selfAttach" seem to work out of the box.

The very old workaround described at https://stackoverflow.com/questions/25438983/com-sun-tools-attach-attachnotsupportedexception-unable-to-open-socket-file-ta/25439003#25439003 seems to help.