#57 Fail if query has multiple statements
For some reason using `.single().flatMapMany()` seems to cause the Flux to get stuck sometimes, perhaps related to batching. This seems to work, although it's not really failing "fast" since it still has to execute the queries and it may have partially inserted some data for the original statement. Nonetheless, it does work sufficiently for now.