Refactor integration test for tx compatibility with Micronaut 3.7
It seems all the transaction managers here confuse @MicronautTest and we need to ensure that
when using Flyway to migrate we tell it specifically the transaction managers to use. We don't
actually need transactions here, but can't find a way to disable it, or clean way to inject
the raw datasource without the transaction managing proxy.