Add validation to prevent bad reactor tools version in images
It would be better to be able to do this by delaying the jib configuration somehow so the correct
version could be determined after dependencies are resolved, but that's greater than my Gradle
capabilities right now. Instead, let's validate the expected version and fail if it doesn't match
so at least we don't build a bad container image that won't work correctly.